Subset a BAM file based on specific tags

This function subselects reads from a BAM file based on the provided tags and writes the subselected reads into separate output BAM files. The subselection is done using a specified tag (default "CB"). This function can utilize multiple cores to perform the operation more efficiently.

Usage

subsetbam(
  inputbam,
  tags,
  outputbams,
  prefixes = NULL,
  TAG = "CB",
  cores = 1,
  verbose = F
)

Arguments

inputbam

A character string specifying the path to the input BAM file.

tags

A character vector of tags to filter the reads by.

outputbams

A character vector specifying the paths to the output BAM files.

prefixes

Optional; a character vector of prefixes to add to the read names in the output BAM files. If NULL, no prefixes are added. Defaults to NULL.

TAG

Optional; the tag used for subselection, default is "CB".

cores

Optional; the number of cores to use for the process, default is 1.

verbose

Option to increase verbosity of output

Value

No return value, called for side effects.

Details

It's important that the length of tags is equal to the length of outputbams. If prefixes is provided, its length must also match the length of outputbams. If any of these conditions is not met, the function will stop and throw an error.

Examples

subsetbam(inputbam = "path/to/input.bam", 
          tags = c("TAG1", "TAG2"), 
          outputbams = c("path/to/output1.bam", "path/to/output2.bam"),
          prefixes = c("prefix1", "prefix2"),
          TAG = "CB",
          cores = 2)
#> File not found:
#> 	path/to/input.bam